+++ src/visit.rs
@@ -8,16 +8,19 @@
 //! e.g. for cargo they are identified from the targets. The tree walker then
 //! follows `mod` statements to recursively visit other referenced files.
 
-use std::borrow::Cow;
 use std::collections::VecDeque;
 use std::sync::Arc;
 
 use anyhow::Context;
 use itertools::Itertools;
-use quote::ToTokens;
+use proc_macro2::{Delimiter, TokenStream, TokenTree};
+use quote::{quote, ToTokens};
 use syn::ext::IdentExt;
 use syn::visit::Visit;
-use syn::{Attribute, ItemFn};
+use syn::{
+    AngleBracketedGenericArguments, Attribute, Expr, GenericArgument, ItemFn, Path, PathArguments,
+    ReturnType, Type, TypeArray, TypeTuple,
+};
 use tracing::{debug, debug_span, trace, trace_span, warn};
 
 use crate::path::TreeRelativePathBuf;
@@ -37,13 +40,18 @@ pub struct Discovered {
 ///
 /// The list of source files includes even those with no mutants.
 pub fn walk_tree(tool: &dyn Tool, root: &Utf8Path, options: &Options) -> Result<Discovered> {
+    let error_exprs = options
+        .error_values
+        .iter()
+        .map(|e| syn::parse_str(e).with_context(|| format!("Failed to parse error value {e:?}")))
+        .collect::<Result<Vec<Expr>>>()?;
     let mut mutants = Vec::new();
     let mut files: Vec<Arc<SourceFile>> = Vec::new();
-
     let mut file_queue: VecDeque<Arc<SourceFile>> = tool.root_files(root)?.into();
     while let Some(source_file) = file_queue.pop_front() {
         check_interrupted()?;
-        let (mut file_mutants, more_files) = walk_file(root, Arc::clone(&source_file), options)?;
+        let (mut file_mutants, more_files) =
+            walk_file(root, Arc::clone(&source_file), options, &error_exprs)?;
         // We'll still walk down through files that don't match globs, so that
         // we have a chance to find modules underneath them. However, we won't
         // collect any mutants from them, and they don't count as "seen" for
@@ -87,18 +95,20 @@ fn walk_file(
     root: &Utf8Path,
     source_file: Arc<SourceFile>,
     options: &Options,
+    error_exprs: &[Expr],
 ) -> Result<(Vec<Mutant>, Vec<TreeRelativePathBuf>)> {
     let _span = debug_span!("source_file", path = source_file.tree_relative_slashes()).entered();
     debug!("visit source file");
     let syn_file = syn::parse_str::<syn::File>(&source_file.code)
         .with_context(|| format!("failed to parse {}", source_file.tree_relative_slashes()))?;
     let mut visitor = DiscoveryVisitor {
-        root: root.to_owned(),
-        source_file,
+        error_exprs,
         more_files: Vec::new(),
         mutants: Vec::new(),
         namespace_stack: Vec::new(),
         options,
+        root: root.to_owned(),
+        source_file,
     };
     visitor.visit_file(&syn_file);
     Ok((visitor.mutants, visitor.more_files))
@@ -123,20 +133,24 @@ struct DiscoveryVisitor<'o> {
     more_files: Vec<TreeRelativePathBuf>,
 
     /// Global options.
+    #[allow(unused)]
     options: &'o Options,
+
+    /// Parsed error expressions, from the config file or command line.
+    error_exprs: &'o [Expr],
 }
 
 impl<'o> DiscoveryVisitor<'o> {
-    fn collect_fn_mutants(&mut self, return_type: &syn::ReturnType, span: &proc_macro2::Span) {
+    fn collect_fn_mutants(&mut self, return_type: &ReturnType, span: &proc_macro2::Span) {
         let full_function_name = Arc::new(self.namespace_stack.join("::"));
         let return_type_str = Arc::new(return_type_to_string(return_type));
-        let mut new_mutants = return_value_replacements(return_type, self.options)
+        let mut new_mutants = return_type_replacements(return_type, self.error_exprs)
             .into_iter()
-            .map(|replacement| Mutant {
-                function_name: Arc::clone(&full_function_name),
-                replacement,
-                return_type: Arc::clone(&return_type_str),
+            .map(|rep| Mutant {
                 source_file: Arc::clone(&self.source_file),
+                function_name: Arc::clone(&full_function_name),
+                return_type: Arc::clone(&return_type_str),
+                replacement: tokens_to_pretty_string(&rep),
                 span: span.into(),
                 genre: Genre::FnValue,
             })
@@ -159,8 +173,7 @@ impl<'o> DiscoveryVisitor<'o> {
     where
         F: FnOnce(&mut Self) -> T,
     {
-        let name = remove_excess_spaces(name);
-        self.namespace_stack.push(name.clone());
+        self.namespace_stack.push(name.to_owned());
         let r = f(self);
         assert_eq!(self.namespace_stack.pop().unwrap(), name);
         r
@@ -170,8 +183,7 @@ impl<'o> DiscoveryVisitor<'o> {
 impl<'ast> Visit<'ast> for DiscoveryVisitor<'_> {
     /// Visit top-level `fn foo()`.
     fn visit_item_fn(&mut self, i: &'ast ItemFn) {
-        // TODO: Filter out more inapplicable fns.
-        let function_name = remove_excess_spaces(&i.sig.ident.to_token_stream().to_string());
+        let function_name = tokens_to_pretty_string(&i.sig.ident);
         let _span = trace_span!(
             "fn",
             line = i.sig.fn_token.span.start().line,
@@ -191,7 +203,7 @@ impl<'ast> Visit<'ast> for DiscoveryVisitor<'_> {
     fn visit_impl_item_fn(&mut self, i: &'ast syn::ImplItemFn) {
         // Don't look inside constructors (called "new") because there's often no good
         // alternative.
-        let function_name = remove_excess_spaces(&i.sig.ident.to_token_stream().to_string());
+        let function_name = tokens_to_pretty_string(&i.sig.ident);
         let _span = trace_span!(
             "fn",
             line = i.sig.fn_token.span.start().line,
@@ -216,25 +228,17 @@ impl<'ast> Visit<'ast> for DiscoveryVisitor<'_> {
         if attrs_excluded(&i.attrs) {
             return;
         }
-        let type_name = type_name_string(&i.self_ty);
+        let type_name = tokens_to_pretty_string(&i.self_ty);
         let name = if let Some((_, trait_path, _)) = &i.trait_ {
             let trait_name = &trait_path.segments.last().unwrap().ident;
             if trait_name == "Default" {
-                // We don't know (yet) how to generate an interestingly-broken
-                // Default::default.
+                // Can't think of how to generate a viable different default.
                 return;
             }
-            format!(
-                "<impl {} for {}>",
-                trait_name,
-                remove_excess_spaces(&type_name)
-            )
+            format!("<impl {trait_name} for {type_name}>")
         } else {
             type_name
         };
-        // Make an approximately-right namespace.
-        // TODO: For `impl X for Y` get both X and Y onto the namespace
-        // stack so that we can show a more descriptive name.
         self.in_namespace(&name, |v| syn::visit::visit_item_impl(v, i));
     }
 
@@ -303,130 +307,300 @@ impl<'ast> Visit<'ast> for DiscoveryVisitor<'_> {
 }
 
 /// Generate replacement text for a function based on its return type.
-fn return_value_replacements(
-    return_type: &syn::ReturnType,
-    options: &Options,
-) -> Vec<Cow<'static, str>> {
-    let mut reps: Vec<Cow<'static, str>> = Vec::new();
+fn return_type_replacements(return_type: &ReturnType, error_exprs: &[Expr]) -> Vec<TokenStream> {
     match return_type {
-        syn::ReturnType::Default => reps.push(Cow::Borrowed("()")),
-        syn::ReturnType::Type(_rarrow, box_typ) => match &**box_typ {
-            syn::Type::Never(_) => {
-                // In theory we could mutate this to a function that just
-                // loops or sleeps, but it seems unlikely to be useful,
-                // so generate nothing.
-            }
-            syn::Type::Path(syn::TypePath { path, .. }) => {
-                // dbg!(&path);
-                if path.is_ident("bool") {
-                    reps.push("true".into());
-                    reps.push("false".into());
-                } else if path.is_ident("String") {
-                    // TODO: Detect &str etc.
-                    reps.push(r#"String::new()"#.into());
-                    reps.push(r#""xyzzy".into()"#.into());
-                } else if path_is_result(path) {
-                    // TODO: Try this for any path ending in `Result`, to handle e.g. `crate::Result`.
-                    // TODO: Recursively generate for types inside the Ok side of the Result.
-                    reps.push("Ok(Default::default())".into());
+        ReturnType::Default => vec![quote! { () }],
+        ReturnType::Type(_rarrow, type_) => type_replacements(type_, error_exprs),
+    }
+}
+
+/// Generate some values that we hope are reasonable replacements for a type.
+///
+/// This is really the heart of cargo-mutants.
+fn type_replacements(type_: &Type, error_exprs: &[Expr]) -> Vec<TokenStream> {
+    let mut reps = Vec::new();
+    match type_ {
+        Type::Path(syn::TypePath { path, .. }) => {
+            // dbg!(&path);
+            if path.is_ident("bool") {
+                reps.push(quote! { true });
+                reps.push(quote! { false });
+            } else if path.is_ident("String") {
+                reps.push(quote! { String::new() });
+                reps.push(quote! { "xyzzy".into() });
+            } else if path_is_unsigned(path) {
+                reps.push(quote! { 0 });
+                reps.push(quote! { 1 });
+            } else if path_is_signed(path) {
+                reps.push(quote! { 0 });
+                reps.push(quote! { 1 });
+                reps.push(quote! { -1 });
+            } else if path_is_nonzero_signed(path) {
+                reps.extend([quote! { 1 }, quote! { -1 }]);
+            } else if path_is_nonzero_unsigned(path) {
+                reps.push(quote! { 1 });
+            } else if path_is_float(path) {
+                reps.push(quote! { 0.0 });
+                reps.push(quote! { 1.0 });
+                reps.push(quote! { -1.0 });
+            } else if path_ends_with(path, "Result") {
+                if let Some(ok_type) = result_ok_type(path) {
                     reps.extend(
-                        options
-                            .error_values
-                            .iter()
-                            .map(|s| format!("Err({})", s).into()),
+                        type_replacements(ok_type, error_exprs)
+                            .into_iter()
+                            .map(|rep| {
+                                quote! { Ok(#rep) }
+                            }),
                     );
                 } else {
-                    reps.push("Default::default()".into());
+                    // A result but with no type arguments, like `fmt::Result`; hopefully
+                    // the Ok value can be constructed with Default.
+                    reps.push(quote! { Ok(Default::default()) });
                 }
+                reps.extend(error_exprs.iter().map(|error_expr| {
+                    quote! { Err(#error_expr) }
+                }));
+            } else if let Some(boxed_type) = match_first_type_arg(path, "Box") {
+                reps.extend(
+                    type_replacements(boxed_type, error_exprs)
+                        .into_iter()
+                        .map(|rep| {
+                            quote! { Box::new(#rep) }
+                        }),
+                )
+            } else if let Some(some_type) = match_first_type_arg(path, "Option") {
+                reps.push(quote! { None });
+                reps.extend(
+                    type_replacements(some_type, error_exprs)
+                        .into_iter()
+                        .map(|rep| {
+                            quote! { Some(#rep) }
+                        }),
+                );
+            } else if let Some(boxed_type) = match_first_type_arg(path, "Vec") {
+                // Generate an empty Vec, and then a one-element vec for every recursive
+                // value.
+                reps.push(quote! { vec![] });
+                reps.extend(
+                    type_replacements(boxed_type, error_exprs)
+                        .into_iter()
+                        .map(|rep| {
+                            quote! { vec![#rep] }
+                        }),
+                )
+            } else if let Some(inner_type) = match_first_type_arg(path, "Arc") {
+                // TODO: Ideally we should use the path without relying on it being
+                // imported, but we must strip or rewrite the arguments, so that
+                // `std::sync::Arc<String>` becomes either `std::sync::Arc::<String>::new`
+                // or at least `std::sync::Arc::new`. Similarly for other types.
+                reps.extend(
+                    type_replacements(inner_type, error_exprs)
+                        .into_iter()
+                        .map(|rep| {
+                            quote! { Arc::new(#rep) }
+                        }),
+                )
+            } else {
+                reps.push(quote! { Default::default() });
             }
-            syn::Type::Reference(syn::TypeReference {
-                mutability: None,
-                elem,
-                ..
-            }) => match &**elem {
-                // needs a separate `match` because of the box.
-                syn::Type::Path(path) if path.path.is_ident("str") => {
-                    reps.push("\"\"".into());
-                    reps.push("\"xyzzy\"".into());
-                }
-                _ => {
-                    trace!(?box_typ, "Return type is not recognized, trying Default");
-                    reps.push("Default::default()".into());
-                }
-            },
-            syn::Type::Reference(syn::TypeReference {
-                mutability: Some(_),
-                ..
-            }) => {
-                reps.push("Box::leak(Box::new(Default::default()))".into());
+        }
+        Type::Array(TypeArray { elem, len, .. }) => reps.extend(
+            // Generate arrays that repeat each replacement value however many times.
+            // In principle we could generate combinations, but that might get very
+            // large, and values like "all zeros" and "all ones" seem likely to catch
+            // lots of things.
+            type_replacements(elem, error_exprs)
+                .into_iter()
+                .map(|r| quote! { [ #r; #len ] }),
+        ),
+        Type::Reference(syn::TypeReference {
+            mutability: None,
+            elem,
+            ..
+        }) => match &**elem {
+            Type::Path(path) if path.path.is_ident("str") => {
+                reps.push(quote! { "" });
+                reps.push(quote! { "xyzzy" });
             }
             _ => {
-                trace!(?box_typ, "Return type is not recognized, trying Default");
-                reps.push("Default::default()".into());
+                reps.extend(type_replacements(elem, error_exprs).into_iter().map(|rep| {
+                    quote! { &#rep }
+                }));
             }
         },
+        Type::Reference(syn::TypeReference {
+            mutability: Some(_),
+            elem,
+            ..
+        }) => {
+            // Make &mut with static lifetime by leaking them on the heap.
+            reps.extend(type_replacements(elem, error_exprs).into_iter().map(|rep| {
+                quote! { Box::leak(Box::new(#rep)) }
+            }));
+        }
+        Type::Tuple(TypeTuple { elems, .. }) if elems.is_empty() => {
+            reps.push(quote! { () });
+            // TODO: Also recurse into non-empty tuples.
+        }
+        Type::Never(_) => {
+            // In theory we could mutate this to a function that just
+            // loops or sleeps, but it seems unlikely to be useful,
+            // so generate nothing.
+        }
+        _ => {
+            trace!(?type_, "Return type is not recognized, trying Default");
+            reps.push(quote! { Default::default() });
+        }
     }
     reps
 }
 
-fn type_name_string(ty: &syn::Type) -> String {
-    ty.to_token_stream().to_string()
-}
-
-fn return_type_to_string(return_type: &syn::ReturnType) -> String {
+fn return_type_to_string(return_type: &ReturnType) -> String {
     match return_type {
-        syn::ReturnType::Default => String::new(),
-        syn::ReturnType::Type(arrow, typ) => {
+        ReturnType::Default => String::new(),
+        ReturnType::Type(arrow, typ) => {
             format!(
                 "{} {}",
                 arrow.to_token_stream(),
-                remove_excess_spaces(&typ.to_token_stream().to_string())
+                tokens_to_pretty_string(typ)
             )
         }
     }
 }
 
-/// Convert a TokenStream representing a type to a String with typical Rust
-/// spacing between tokens.
+fn path_ends_with(path: &Path, ident: &str) -> bool {
+    path.segments.last().map_or(false, |s| s.ident == ident)
+}
+
+fn path_is_float(path: &Path) -> bool {
+    ["f32", "f64"].iter().any(|s| path.is_ident(s))
+}
+
+fn path_is_unsigned(path: &Path) -> bool {
+    ["u8", "u16", "u32", "u64", "u128", "usize"]
+        .iter()
+        .any(|s| path.is_ident(s))
+}
+
+fn path_is_signed(path: &Path) -> bool {
+    ["i8", "i16", "i32", "i64", "i128", "isize"]
+        .iter()
+        .any(|s| path.is_ident(s))
+}
+
+fn path_is_nonzero_signed(path: &Path) -> bool {
+    if let Some(l) = path.segments.last().map(|p| p.ident.to_string()) {
+        matches!(
+            l.as_str(),
+            "NonZeroIsize"
+                | "NonZeroI8"
+                | "NonZeroI16"
+                | "NonZeroI32"
+                | "NonZeroI64"
+                | "NonZeroI128",
+        )
+    } else {
+        false
+    }
+}
+
+fn path_is_nonzero_unsigned(path: &Path) -> bool {
+    if let Some(l) = path.segments.last().map(|p| p.ident.to_string()) {
+        matches!(
+            l.as_str(),
+            "NonZeroUsize"
+                | "NonZeroU8"
+                | "NonZeroU16"
+                | "NonZeroU32"
+                | "NonZeroU64"
+                | "NonZeroU128",
+        )
+    } else {
+        false
+    }
+}
+
+/// Convert a TokenStream representing some code to a reasonably formatted
+/// string of Rust code.
 ///
-/// This shrinks for example "& 'static" to just "&'static".
-fn remove_excess_spaces(s: &str) -> String {
-    // Walk through looking at space characters, and consider whether we can drop them
-    // without it being ambiguous.
-    //
-    // This is a bit hacky but seems to give reasonably legible results on
-    // typical trees...
-    //
-    // We could instead perhaps do this on the type enum.
-    let mut r = String::with_capacity(s.len());
-    for c in s.chars() {
-        match c {
-            ' ' if r.ends_with("->") => {}
-            ' ' => {
-                // drop spaces following any of these chars
-                if let Some(a) = r.chars().next_back() {
-                    match a {
-                        ':' | '&' | '<' | '>' => continue, // drop the space
-                        _ => {}
+/// [TokenStream] has a `to_string`, but it adds spaces in places that don't
+/// look idiomatic, so this reimplements it in a way that looks better.
+///
+/// This is probably not correctly formatted for all Rust syntax, and only tries
+/// to cover cases that can emerge from the code we generate.
+fn tokens_to_pretty_string<T: ToTokens>(t: T) -> String {
+    use TokenTree::*;
+    let mut b = String::with_capacity(200);
+    let mut ts = t.to_token_stream().into_iter().peekable();
+    while let Some(tt) = ts.next() {
+        match tt {
+            Punct(p) => {
+                let pc = p.as_char();
+                b.push(pc);
+                if ts.peek().is_some() && (b.ends_with("->") || pc == ',' || pc == ';') {
+                    b.push(' ');
+                }
+            }
+            Ident(_) | Literal(_) => {
+                match tt {
+                    Literal(l) => b.push_str(&l.to_string()),
+                    Ident(i) => b.push_str(&i.to_string()),
+                    _ => unreachable!(),
+                };
+                if let Some(next) = ts.peek() {
+                    match next {
+                        Ident(_) | Literal(_) => b.push(' '),
+                        Punct(p) => match p.as_char() {
+                            ',' | ';' | '<' | '>' | ':' | '.' | '!' => (),
+                            _ => b.push(' '),
+                        },
+                        Group(_) => (),
                     }
                 }
             }
-            ':' | ',' | '<' | '>' if r.ends_with(' ') => {
-                // drop spaces preceding these chars
-                r.pop();
+            Group(g) => {
+                match g.delimiter() {
+                    Delimiter::Brace => b.push('{'),
+                    Delimiter::Bracket => b.push('['),
+                    Delimiter::Parenthesis => b.push('('),
+                    Delimiter::None => (),
+                }
+                b.push_str(&tokens_to_pretty_string(g.stream()));
+                match g.delimiter() {
+                    Delimiter::Brace => b.push('}'),
+                    Delimiter::Bracket => b.push(']'),
+                    Delimiter::Parenthesis => b.push(')'),
+                    Delimiter::None => (),
+                }
             }
-            _ => {}
         }
-        r.push(c)
     }
-    r
+    debug_assert!(
+        !b.ends_with(' '),
+        "generated a trailing space: ts={ts:?}, b={b:?}",
+        ts = t.to_token_stream(),
+    );
+    b
 }
 
-fn path_is_result(path: &syn::Path) -> bool {
-    path.segments
-        .last()
-        .map(|segment| segment.ident == "Result")
-        .unwrap_or_default()
+/// If this looks like `Result<T, E>` (optionally with `Result` in some module), return `T`.
+fn result_ok_type(path: &Path) -> Option<&Type> {
+    match_first_type_arg(path, "Result")
+}
+
+/// If this is a path ending in `expected_ident`, return the first type argument.
+fn match_first_type_arg<'p>(path: &'p Path, expected_ident: &str) -> Option<&'p Type> {
+    let last = path.segments.last()?;
+    if last.ident == expected_ident {
+        if let PathArguments::AngleBracketed(AngleBracketedGenericArguments { args, .. }) =
+            &last.arguments
+        {
+            if let Some(GenericArgument::Type(ok_type)) = args.first() {
+                return Some(ok_type);
+            }
+        }
+    }
+    None
 }
 
 /// True if the signature of a function is such that it should be excluded.
@@ -509,46 +683,210 @@ fn attr_is_mutants_skip(attr: &Attribute) -> bool {
     }
     skip
 }
-//     fn path_is_mutants_skip(path: &syn::Path) -> bool {
-//         path.segments
-//             .iter()
-//             .map(|ps| &ps.ident)
-//             .eq(["mutants", "skip"].iter())
-//     }
-
-//     fn list_is_mutants_skip(meta_list: &syn::MetaList) -> bool {
-//         return meta_list.nested.iter().any(|n| match n {
-//             syn::NestedMeta::Meta(syn::Meta::Path(path)) => path_is_mutants_skip(path),
-//             syn::NestedMeta::Meta(syn::Meta::List(list)) => list_is_mutants_skip(list),
-//             _ => false,
-//         });
-//     }
-
-//     if path_is_mutants_skip(&attr.path) {
-//         return true;
-//     }
-
-//     if let Ok(syn::Meta::List(meta_list)) = attr.parse_meta() {
-//         return list_is_mutants_skip(&meta_list);
-//     }
-
-//     false
-// }
 
 #[cfg(test)]
 mod test {
+    use quote::quote;
+    use syn::{parse_quote, Expr, ReturnType};
+
+    use super::{return_type_replacements, tokens_to_pretty_string};
+
     #[test]
     fn path_is_result() {
         let path: syn::Path = syn::parse_quote! { Result<(), ()> };
-        assert!(super::path_is_result(&path));
+        assert!(super::result_ok_type(&path).is_some());
     }
 
     #[test]
-    fn remove_excess_spaces() {
-        use super::remove_excess_spaces as rem;
+    fn pretty_format() {
+        assert_eq!(
+            tokens_to_pretty_string(quote! {
+                <impl Iterator for MergeTrees < AE , BE , AIT , BIT > > :: next
+                -> Option < Self ::  Item >
+            }),
+            "<impl Iterator for MergeTrees<AE, BE, AIT, BIT>>::next -> Option<Self::Item>"
+        );
+        assert_eq!(
+            tokens_to_pretty_string(quote! { Lex < 'buf >::take }),
+            "Lex<'buf>::take"
+        );
+    }
 
-        assert_eq!(rem("<impl Iterator for MergeTrees < AE , BE , AIT , BIT > > :: next -> Option < Self ::  Item >"),
-    "<impl Iterator for MergeTrees<AE, BE, AIT, BIT>>::next -> Option<Self::Item>");
-        assert_eq!(rem("Lex < 'buf >::take"), "Lex<'buf>::take");
+    #[test]
+    fn recurse_into_result_bool() {
+        let return_type: syn::ReturnType = parse_quote! {-> std::result::Result<bool> };
+        let reps = return_type_replacements(&return_type, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["Ok(true)", "Ok(false)",]
+        );
+    }
+
+    #[test]
+    fn recurse_into_result_result_bool() {
+        let return_type: syn::ReturnType = parse_quote! {-> std::result::Result<Result<bool>> };
+        let error_expr: syn::Expr = parse_quote! { anyhow!("mutated") };
+        let reps = return_type_replacements(&return_type, &[error_expr]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &[
+                "Ok(Ok(true))",
+                "Ok(Ok(false))",
+                "Ok(Err(anyhow!(\"mutated\")))",
+                "Err(anyhow!(\"mutated\"))"
+            ]
+        );
+    }
+
+    #[test]
+    fn u16_replacements() {
+        let reps = return_type_replacements(&parse_quote! { -> u16 }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["0", "1",]
+        );
+    }
+
+    #[test]
+    fn isize_replacements() {
+        let reps = return_type_replacements(&parse_quote! { -> isize }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["0", "1", "-1"]
+        );
+    }
+
+    #[test]
+    fn nonzero_integer_replacements() {
+        let reps = return_type_replacements(&parse_quote! { -> std::num::NonZeroIsize }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["1", "-1"]
+        );
+
+        let reps = return_type_replacements(&parse_quote! { -> std::num::NonZeroUsize }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["1"]
+        );
+
+        let reps = return_type_replacements(&parse_quote! { -> std::num::NonZeroU32 }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["1"]
+        );
+    }
+
+    #[test]
+    fn unit_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> () }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["()"]
+        );
+    }
+
+    #[test]
+    fn result_unit_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> Result<(), Error> }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["Ok(())"]
+        );
+
+        let reps = return_type_replacements(&parse_quote! { -> Result<()> }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["Ok(())"]
+        );
+    }
+
+    #[test]
+    fn option_usize_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> Option<usize> }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["None", "Some(0)", "Some(1)"]
+        );
+    }
+
+    #[test]
+    fn box_usize_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> Box<usize> }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["Box::new(0)", "Box::new(1)"]
+        );
+    }
+
+    #[test]
+    fn box_unrecognized_type_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> Box<MyObject> }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["Box::new(Default::default())"]
+        );
+    }
+
+    #[test]
+    fn vec_string_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> std::vec::Vec<String> }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["vec![]", "vec![String::new()]", "vec![\"xyzzy\".into()]"]
+        );
+    }
+
+    #[test]
+    fn float_replacement() {
+        let reps = return_type_replacements(&parse_quote! { -> f32 }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["0.0", "1.0", "-1.0"]
+        );
+    }
+
+    #[test]
+    fn ref_replacement_recurses() {
+        let reps = return_type_replacements(&parse_quote! { -> &bool }, &[]);
+        assert_eq!(
+            reps.iter().map(tokens_to_pretty_string).collect::<Vec<_>>(),
+            &["&true", "&false"]
+        );
+    }
+
+    #[test]
+    fn array_replacement() {
+        assert_eq!(
+            replace(&parse_quote! { -> [u8; 256] }, &[]),
+            &["[0; 256]", "[1; 256]"]
+        );
+    }
+
+    #[test]
+    fn arc_replacement() {
+        // Also checks that it matches the path, even using an atypical path.
+        // TODO: Ideally this would be fully qualified like `alloc::sync::Arc::new(String::new())`.
+        assert_eq!(
+            replace(&parse_quote! { -> alloc::sync::Arc<String> }, &[]),
+            &["Arc::new(String::new())", "Arc::new(\"xyzzy\".into())"]
+        );
+    }
+
+    // #[test]
+    // fn rc_replacement() {
+    //     // Also checks that it matches the path, even using an atypical path.
+    //     // TODO: Ideally this would be fully qualified like `alloc::sync::Rc::new(String::new())`.
+    //     assert_eq!(
+    //         replace(&parse_quote! { -> alloc::sync::Rc<String> }, &[]),
+    //         &["Rc::new(String::new())", "Rc::new(\"xyzzy\".into())"]
+    //     );
+    // }
+
+    fn replace(return_type: &ReturnType, error_exprs: &[Expr]) -> Vec<String> {
+        return_type_replacements(return_type, error_exprs)
+            .into_iter()
+            .map(tokens_to_pretty_string)
+            .collect::<Vec<_>>()
     }
 }