+++ testdata/tree/well_tested/src/lib.rs
@@ -8,11 +8,13 @@
 
 #![allow(unused, dead_code)]
 
+mod arc;
 mod empty_fns;
 mod inside_mod;
 mod item_mod;
 mod methods;
 mod nested_function;
+mod numbers;
 mod result;
 pub mod simple_fns;
 mod struct_with_lifetime;